requirements.txt
autopep8
dnspython
parametrize_from_file
pycodestyle
pylint
pytest
requests
rope
